requirements.txt
typer[all]
pydantic
pyyaml
